requirements.txt
altgraph==0.17.5
asgiref==3.11.1
bleach==6.3.0
crispy-tailwind==1.0.3
Django==6.0.3
django-crispy-forms==2.6
django-htmx==1.27.0
django-markdownx==4.0.9
django-tailwind==4.4.2
Markdown==3.10.2
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pytailwindcss==0.3.0
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2025.3
waitress==3.0.2
webencodings==0.5.1
